Attempting to place consecutive non-Markdown underscores into text
Markdown appears to interpret a series of underscores as a separation line.
But underscores are also used by CardIAC-B to show empty values. So showing
an example with lots of empty values is interpreted by Markdown as separation
lines. This change inserts HTML non-breaking spaces ( &nbsp; ) for spaces in
lines of CardIAC-B code.